refactor: move junction util to AsstPlatformWin32
